feat: total rework of job management, allow cancel
The previous threading was very manually handled. We now use asyncio
completely, which will mean that CPU intensive async code will all
run on the same thread but this is not an issue since the majority
of async code is IO and so will be run on separate threads by asyncio

The advantange of this approach is simplicity. We now have a fully
working cancelling system to allow for cancelling of long running
tests, and perhaps other operations if required in future.