feat: run file as whole (#36)
Currently the plugin runs all tests individually, spawning a separate
process for each one. This change allows a file to be run as whole, as
long as the output parser knows how to parse the runner's output.

As a consequence of this, we're also now parsing the namespaces in which
a test is defined, similarly to vim-test where it uses the indents.
Currently they are only used in parsing output, but they will be able to
be run as groups as well soon enough.