feat: namespace/file running and result tracking
A number of large changes happening here. Mostly focused on how the
plugin tracks the running processes. The process manager and handler
have both been simplified with the test running logic separated to a
separate class.

As a benefit we can now show the file and namespaces as
running/passed/failed in the summary window and parse the output for
namespaces from external running (i.e. nvim-dap)