feat: namespace support (#37)
BREAKING CHANGE: Some more advanced functions have changed. They are
not likely being used by anyone but since so much has changed in this
PR I'm being safe and calling it out that integrations may break.
Also the defaults for icons have changed.

A number of large changes happening here. Mostly focused on how the
plugin tracks the running processes. The process manager and handler
have both been simplified with the test running logic separated to a
separate class.

The main feature of this PR is that vim-ultest is now parsing files as
structured trees and using namespaces.  This allows the plugin to have
a much better understanding of the code structure.  As a primary
benefit we can now show the file and namespaces as
running/passed/failed in the summary window and parse the output for
namespaces from external running (i.e. nvim-dap). Runners that do not
have output parsing support will not show namspaces/files as
running/passed/failed.