Clarify the sentence about Neovim 0.4.4 support
This might seem like pointless nitpicking, but someone on Matrix actually read this sentence in the README to mean "Once Neovim 0.5 is released, support for all versions >0.4.4 (i.e., including 0.5) will be removed". This doesn't seem to me to be a meaningful interpretation, but it's true that the sentence is ambiguous if you read it literally. Hence this patch.